20.10.16

+++ src/main/java/ru/settletale/client/opengl/VertexInfo.java
@@ -3,9 +3,9 @@ package ru.settletale.client.opengl;
 public class VertexInfo {
 	public static final int SIZE = Double.BYTES * 3 + Float.BYTES * 5 + Byte.BYTES * 3;
 	
-	public double pX;
-	public double pY;
-	public double pZ;
+	public float pX;
+	public float pY;
+	public float pZ;
 	public float tX;
 	public float tY;
 	public float nX;
